chore: inline ComponentsViewModel into PaywallViewModel
Single implementer on this branch — introduce the shared interface when
the other implementer actually lands.
